Some corrections.
